working on unit crud

+++ app/Http/Controllers/Auth/UserController.php
@@ -9,6 +9,7 @@ use Illuminate\Contracts\View\View;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
@@ -27,6 +28,8 @@ class UserController extends Controller
      */
     public function index(): View
     {
+        Gate::authorize('app.user.index');
+
         return view('pages.users.index', $this->services->accessPermissions());
     }
 
@@ -37,6 +40,8 @@ class UserController extends Controller
      */
     public function getUsers(): LengthAwarePaginator
     {
+        Gate::authorize('app.user.index');
+
         return $this->services->getUsers();
     }
 
@@ -45,6 +50,8 @@ class UserController extends Controller
      */
     public function getRoles(): Collection|array
     {
+        Gate::authorize('app.user.index');
+
         return $this->services->getRoles();
     }
 
@@ -53,6 +60,8 @@ class UserController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        Gate::authorize('app.user.create');
+
         return $this->services->validateStore($request)->store($request);
     }
 
@@ -69,6 +78,8 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id): JsonResponse
     {
+        Gate::authorize('app.user.edit');
+
        return $this->services->validateUpdate($request, $id)->update($request, $id);
     }
 
@@ -77,6 +88,8 @@ class UserController extends Controller
      */
     public function destroy(string $id): JsonResponse
     {
+        Gate::authorize('app.user.delete');
+
         return $this->services->delete($id);
     }
 }
